fichier et bug

+++ Seuillage.py
@@ -7,7 +7,7 @@ import Outils
 #Prend en paramètre une image de couleur et renvoie un tableau représentant une image de gris
 def grayscale(picture):
     pictureX , pictureY, profondeur = picture.shape
-    result = np.zeros(((pictureX,pictureY,1)))
+    result =  np.uint8(np.zeros(((pictureX,pictureY,1))))
     #formule de luminosité Luminance = 0,2126 × Rouge + 0,7152 × Vert + 0,0722 × Bleu
     #attention open CV charge en BGR et non RGB
     for i in range(pictureX):
@@ -18,7 +18,7 @@ def grayscale(picture):
 #Prend en paramètre une image de couleur et un seuil puis binarise en fonction du seuil
 def binarisation(picture,threshold):
     pictureX , pictureY, profondeur = picture.shape
-    result = np.zeros(((pictureX,pictureY,1)))
+    result =  np.uint8(np.zeros(((pictureX,pictureY,1))))
     for i in range(pictureX):
         for j in range(pictureY):
             result[i][j][0] = 1 if picture[i][j][0]>threshold else 0
